+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Post v1 a change is further considered breaking if the state file format is inco
 *Major*
 
 * Added the Necromancer role. ([#211](https://github.com/hjwylde/werewolf/issues/211))
+* Renamed Scapegoat's allowedVoters to chosenVoters. ([#211](https://github.com/hjwylde/werewolf/issues/211))
 
 *Minor*